test/html/parser/input/pending-spec-changes-plaintext-unsafe.dat

#data
<body><table>fillertext
#errors
(1,6): expected-doctype-but-got-start-tag
(1,14): invalid-codepoint
(1,14): invalid-codepoint-in-table-text
(1,21): invalid-codepoint
(1,21): invalid-codepoint-in-table-text
(1,26): invalid-codepoint
(1,26): invalid-codepoint-in-table-text
(1,26): foster-parenting-character-in-table
(1,26): foster-parenting-character-in-table
(1,26): foster-parenting-character-in-table
(1,26): foster-parenting-character-in-table
(1,26): foster-parenting-character-in-table
(1,26): foster-parenting-character-in-table
(1,26): foster-parenting-character-in-table
(1,26): foster-parenting-character-in-table
(1,26): foster-parenting-character-in-table
(1,26): foster-parenting-character-in-table
(1,26): eof-in-table
#new-errors
(1:14) unexpected-null-character
(1:21) unexpected-null-character
(1:26) unexpected-null-character
#document
| <html>
|   <head>
|   <body>
|     "fillertext"
|     <table>
